# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KwunMan OS - 冠文鋼結構管理系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [projects]
      summary: List invoices, optionally filtered by quoteId
      parameters:
        - name: quoteId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [projects]
      summary: Create a new invoice for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /invoices/{invoiceId}:
    patch:
      operationId: updateInvoice
      tags: [projects]
      summary: Update an invoice (e.g. mark as paid)
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [projects]
      summary: Delete an invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PAYMENTS ─────────────────────────────────────────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [projects]
      summary: List payments for a project
      parameters:
        - name: quoteId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [projects]
      summary: Add a payment record to a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}:
    delete:
      operationId: deletePayment
      tags: [projects]
      summary: Delete a payment record
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── QUOTE ITEMS ──────────────────────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: ["pending", "approved", "rejected"]
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{expenseId}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense (e.g. approve/reject)
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EXPENSE CATEGORIES ────────────────────────────────────────────────────────
...
  /note-templates:
    get:
      operationId: listNoteTemplates
      tags: [note-templates]
      summary: List all note templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteTemplate"
    post:
      operationId: createNoteTemplate
      tags: [note-templates]
      summary: Create a note template (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteTemplate"
  /note-templates/{templateId}:
    put:
      operationId: updateNoteTemplate
      tags: [note-templates]
      summary: Update a note template (admin only)
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteTemplateInput"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteTemplate"
    delete:
      operationId: deleteNoteTemplate
      tags: [note-templates]
      summary: Delete a note template (admin only)
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AUTH ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Payment:
      type: object
      required:
        - id
        - quoteId
        - amount
        - paymentType
        - createdAt
      properties:
        id:
          type: integer
        quoteId:
          type: string
        amount:
          type: number
        paymentType:
          type: string
          description: "訂金 / 中期款 / 尾款 / 其他"
        paymentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required:
        - quoteId
        - amount
        - paymentType
      properties:
        quoteId:
          type: string
        amount:
          type: number
        paymentType:
          type: string
        paymentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── QUOTE ITEM ───────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      required:
        - id
        - invoiceNo
        - quoteId
        - issueDate
        - amount
        - description
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        quoteId:
          type: string
        issueDate:
          type: string
        dueDate:
          type: ["string", "null"]
        amount:
          type: number
        description:
          type: string
        status:
          type: string
          enum: ["未收", "已收", "部分收款", "取消"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    InvoiceInput:
      type: object
      required:
        - quoteId
        - issueDate
        - amount
        - description
      properties:
        invoiceNo:
          type: ["string", "null"]
        quoteId:
          type: string
        issueDate:
          type: string
        dueDate:
          type: ["string", "null"]
        amount:
          type: number
        description:
          type: string
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    InvoiceUpdate:
      type: object
      properties:
        issueDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Expense:
      type: object
      required:
        - id
        - projectId
        - description
        - amount
        - status
        - submittedAt
      properties:
        id:
          type: integer
        projectId:
          type: ["string", "null"]
        description:
          type: string
        amount:
          type: number
        category:
          type: ["string", "null"]
        bank:
          type: ["string", "null"]
        merchant:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: ["pending", "approved", "rejected"]
        receiptDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        submittedAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ExpenseInput:
      type: object
      required:
        - description
        - amount
      properties:
        projectId:
          type: ["string", "null"]
        description:
          type: string
        amount:
          type: number
        category:
          type: ["string", "null"]
        bank:
          type: ["string", "null"]
        merchant:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        receiptDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ExpenseUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
          enum: ["pending", "approved", "rejected", null]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        bank:
          type: ["string", "null"]
        merchant:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        receiptDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        projectId:
          type: ["string", "null"]

    # ─── EXPENSE CATEGORIES ────────────────────────────────────────────────────
...
    NoteTemplate:
      type: object
      required: [id, code, title, content, sortOrder, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
    NoteTemplateInput:
      type: object
      required: [code, title, content]
      properties:
        code:
          type: string
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
...